Don't inline methods that could potentially generate ambiguity errors
If we have a call signature that's wide enough to cover one of the ambiguities
of the methods we're inlining, refuse to do the inlining and require a dynamic
dispatch instead. The other option would be to "union split" that signature out,
but that would require a decent amount of extra code and it is not clear how
important that speed up would be given that the signature is already going to
be fairly wide if it intersects an ambiguity (so the dispatch gate might actually
be better). If we encounter a case where that's useful, we can add that feature.
For now, this fixes #30118.